MyGames+ Create Game (ยังไม่เสร็จ)
-รวมหน้า MyGamesเเละCreate Game เป็นหน้าเดียวกัน
-เมิ่อเลื่อนpageviewจนถึงสุดlist จะเป็น ปุ่มเป็น create summary เเละเปลี่ยนเป็นคำอธิบาย
