+++ lib/screens/all.dart
@@ -1,6 +1,5 @@
 export 'home.dart';
 export 'profile.dart';
 export 'settings.dart';
-export 'creategame.dart';
 export 'history.dart';
 export 'mygames.dart';
