+++ lib/screens/mygames.dart
@@ -1,4 +1,5 @@
 import 'package:brainboost/component/colors.dart';
+import 'package:brainboost/screens/creategame.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 
@@ -20,12 +21,14 @@ class _MyGamesState extends State<MyGames> {
     "Software Engineering",
     "Cybersecurity",
     "Data Science",
+    ""
   ];
 
   final List<String> imagePaths = [
     'assets/images/photomain.png',
     'assets/images/photomain2.png',
     'assets/images/photomain3.png',
+    'assets/images/Add.png'
   ];
 
   @override
@@ -66,10 +69,9 @@ class _MyGamesState extends State<MyGames> {
             controller: _scrollController,
             child: Center(
               child: Column(
-                children: [
+                children: <Widget>[
                   Container(
-                    padding:
-                        const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
+                    padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
                     decoration: BoxDecoration(
                       color: AppColors.cardBackground,
                       borderRadius: BorderRadius.circular(50),
@@ -100,8 +102,7 @@ class _MyGamesState extends State<MyGames> {
                   ),
                   const SizedBox(height: 20),
                   Container(
-                    padding:
-                        const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
+                    padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
                     child: Text(
                       titles[_currentPage],
                       style: const TextStyle(
@@ -147,45 +148,104 @@ class _MyGamesState extends State<MyGames> {
                     ),
                   ),
                   const SizedBox(height: 5),
-                  ElevatedButton.icon(
-                    onPressed: () {},
-                    icon: SvgPicture.asset(
-                      'assets/images/game.svg',
-                      width: 35,
-                      height: 35,
+                  if (_currentPage == titles.length - 1)
+                    Column(
+                      children: [
+                        const Text(
+                          "Learn more about Lecture?",
+                          style: TextStyle(
+                            color: Colors.white,
+                            fontSize: 14,
+                            fontStyle: FontStyle.italic,
+                          ),
+                        ),
+                        const SizedBox(height: 20),
+                        ElevatedButton(
+                          onPressed: () {
+                            Navigator.push(
+                              context,
+                              MaterialPageRoute(
+                                builder: (context) => const UploadFileScreen(),
+                              ),
+                            );
+                          },
+                          style: ElevatedButton.styleFrom(
+                            backgroundColor: AppColors.neutralBackground,
+                            shape: RoundedRectangleBorder(
+                              borderRadius: BorderRadius.circular(20),
+                            ),
+                            padding: const EdgeInsets.symmetric(
+                              horizontal: 16,
+                              vertical: 12,
+                            ),
+                          ),
+                          child: Row(
+                            mainAxisSize: MainAxisSize.min,
+                            children: [
+                              SvgPicture.asset(
+                                'assets/images/game.svg',
+                                width: 24,
+                                height: 24,
+                                color: Colors.white,
+                              ),
+                              const SizedBox(width: 8),
+                              const Text(
+                                'Create Summary',
+                                style: TextStyle(
+                                  color: Colors.white,
+                                  fontSize: 16,
+                                  fontWeight: FontWeight.bold,
+                                ),
+                              ),
+                            ],
+                          ),
+                        ),
+                      ],
+                    )
+                  else
+                    Column(
+                      children: [
+                        ElevatedButton.icon(
+                          onPressed: () {},
+                          icon: SvgPicture.asset(
+                            'assets/images/game.svg',
+                            width: 35,
+                            height: 35,
+                          ),
+                          label: const Text(
+                            "Play",
+                            style: TextStyle(
+                              fontSize: 25,
+                              fontWeight: FontWeight.bold,
+                            ),
+                          ),
+                          style: ElevatedButton.styleFrom(
+                            backgroundColor: Colors.white,
+                            foregroundColor: AppColors.buttonText,
+                            padding: const EdgeInsets.symmetric(
+                              horizontal: 25,
+                              vertical: 10,
+                            ),
+                            shape: RoundedRectangleBorder(
+                              borderRadius: BorderRadius.circular(30),
+                            ),
+                            side: BorderSide(
+                              color: AppColors.buttonBorder,
+                              width: 2,
+                            ),
+                          ),
+                        ),
+                        const SizedBox(height: 10),
+                        const Text(
+                          "You have played 2 days ago",
+                          style: TextStyle(
+                            color: Colors.white,
+                            fontSize: 14,
+                            fontStyle: FontStyle.italic,
+                          ),
+                        ),
+                      ],
                     ),
-                    label: const Text(
-                      "Play",
-                      style: TextStyle(
-                        fontSize: 25,
-                        fontWeight: FontWeight.bold,
-                      ),
-                    ),
-                    style: ElevatedButton.styleFrom(
-                      backgroundColor: Colors.white,
-                      foregroundColor: AppColors.buttonText,
-                      padding: const EdgeInsets.symmetric(
-                        horizontal: 25,
-                        vertical: 10,
-                      ),
-                      shape: RoundedRectangleBorder(
-                        borderRadius: BorderRadius.circular(30),
-                      ),
-                      side: BorderSide(
-                        color: AppColors.buttonBorder,
-                        width: 2,
-                      ),
-                    ),
-                  ),
-                  const SizedBox(height: 10),
-                  const Text(
-                    "You have played 2 days ago",
-                    style: TextStyle(
-                      color: Colors.white,
-                      fontSize: 14,
-                      fontStyle: FontStyle.italic,
-                    ),
-                  ),
                   const SizedBox(height: 10),
                   Container(
                     width: double.infinity,
@@ -249,71 +309,58 @@ class _MyGamesState extends State<MyGames> {
               ),
             ),
           ),
-          if (_showButtons)
-            Positioned(
-              bottom: 90,
-              left: 20,
-              right: 20,
-              child: Column(
-                mainAxisSize: MainAxisSize.min,
-                children: [
-                  Row(
-                    mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                    children: [
-                      ElevatedButton(
-                        onPressed: () {},
-                        style: ElevatedButton.styleFrom(
-                          backgroundColor: Colors.white,
-                          foregroundColor: AppColors.buttonText,
-                          shape: RoundedRectangleBorder(
-                            borderRadius: BorderRadius.circular(20),
-                          ),
-                        ),
-                        child: const Text("Re version"),
-                      ),
-                      ElevatedButton(
-                        onPressed: () {
-                          // Handle "Add Lecture"
-                        },
-                        style: ElevatedButton.styleFrom(
-                          backgroundColor: Colors.white,
-                          foregroundColor: AppColors.buttonText,
-                          shape: RoundedRectangleBorder(
-                            borderRadius: BorderRadius.circular(20),
-                          ),
-                        ),
-                        child: const Text("Add Lecture"),
-                      ),
-                    ],
-                  ),
-                  const SizedBox(height: 10),
-                  ElevatedButton.icon(
-                    onPressed: () {},
-                    icon: SvgPicture.asset(
-                      'assets/images/game.svg',
-                      width: 35,
-                      height: 35,
-                    ),
-                    label: const Text(
-                      "Play",
-                      style: TextStyle(
-                        fontSize: 20,
-                        color: AppColors.buttonText,
-                      ),
-                    ),
-                    style: ElevatedButton.styleFrom(
-                      backgroundColor: const Color.fromARGB(255, 240, 239, 240),
-                      shape: RoundedRectangleBorder(
-                        borderRadius: BorderRadius.circular(20),
-                      ),
-                    ),
-                  ),
-                ],
-              ),
-            ),
         ],
       ),
-      // bottomNavigationBar: Navbar(),
+    );
+  }
+}
+
+class IconTitleButton extends StatelessWidget {
+  final String title;
+  final String iconPath;
+  final VoidCallback onPressed;
+
+  const IconTitleButton({
+    required this.title,
+    required this.iconPath,
+    required this.onPressed,
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return ElevatedButton(
+      onPressed: onPressed,
+      style: ElevatedButton.styleFrom(
+        backgroundColor: AppColors.neutralBackground,
+        shape: RoundedRectangleBorder(
+          borderRadius: BorderRadius.circular(20),
+        ),
+        padding: const EdgeInsets.symmetric(
+          horizontal: 16,
+          vertical: 12,
+        ),
+      ),
+      child: Row(
+        mainAxisSize: MainAxisSize.min,
+        children: [
+          SvgPicture.asset(
+            iconPath,
+            width: 24,
+            height: 24,
+            color: Colors.white,
+          ),
+          const SizedBox(width: 8),
+          Text(
+            title,
+            style: const TextStyle(
+              color: Colors.white,
+              fontSize: 16,
+              fontWeight: FontWeight.bold,
+            ),
+          ),
+        ],
+      ),
     );
   }
 }
